--- cooking-ninja/src/App.css
@@ -1,6 +1,0 @@
-.App {
-  min-height: 100%;
-}
-.App.dark {
-  background: #333;
-}
