--- cooking-ninja/src/components/Searchbar.css
@@ -1,7 +1,0 @@
-.searchbar form {
-  display: flex;
-  align-items: center;
-}
-.searchbar input {
-  margin-left: 10px;
-}
