--- cooking-ninja/src/pages/home/Home.css
@@ -1,4 +1,0 @@
-.home {
-  max-width: 1200px;
-  margin: 60px auto;
-}
