+++ my-money/public/manifest.json
